requirements.txt
Django==3.2.10
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2==2.9.3